third_party/WebKit/LayoutTests/fast/dom/custom/lifecycle-created-parser-script-expected.txt

CONSOLE WARNING: line 8: document.registerElement is deprecated and will be removed in M73, around March 2019. Please use window.customElements.define instead. See https://www.chromestatus.com/features/4642138092470272 for more details.
This test ensures that the lifecycle callback of a parser-made element is visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.callbacksCalled is true
PASS successfullyParsed is true

TEST COMPLETE
